kyc_backend/requirements/base.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-decouple==3.8
# psycopg2-binary==2.9.9
Pillow==10.4.0
celery==5.4.0
redis==5.0.8
django-storages==1.14.4
boto3==1.35.0
django-filter==24.3
drf-spectacular==0.27.2
